requirements.txt
tiktoken>=0.12.0
fastapi>=0.135.1
uvicorn>=0.41.0
openai>=2.26.0
pydantic>=2.12.5
python-dotenv>=1.2.2